cdh/storage: ensure block numbers are found before to format the device
In order to format the ephemeral storage device with data integrity, the
luks-encrypt-storage script will read the block numbers to have its pages
cleaned. This introduced a check on the script to ensure that at least one
block is read, otherwise it will bail out as this might indicate a bug on
the logic to find blocks.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>