keyprovider: Pin a specific version of skopeo
The dependency on skopeo is quite fragile as there are several versions
of the project that would generate an encrypted image with a gzip header
that Confidential Containers simply cannot deal with.

For now, let's use f64a376, from Aug 14th 2024 as, empirically, we know
that it works.

We must have tests implemented in order to cover the documentation we
provide, and whether it works or not.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>