cdh/kms: cckbc use AA to get evidence rather than token
When we use one AA and multiple CDH to connect to KBSes in one
confidential VM, AATokenProvider requires the KBS be the same. If we
use AAEvidenceProvider, the different target KBS address can be
specified in CDH's config and share a same AA.

This will also make the workflow simpler. Before this commit, if we want
to get a confidential resource from KBS, AA should connect to a KBS to
get a token. Then CDH get the token from AA. Then CDH access the KBS.

After this commit, only CDH will access the KBS.

This change will require the KBS that does RCAR and stores resource be
the same in CoCo, s.t. background check model. which is up to now
nearly all the deployments we have met.

This will not break the passport model because kbs_protocol crate still
provides a way for developers to integrate either passport model or
background check model in their own code.

Signed-off-by: Xynnn007 <xynnn@linux.alibaba.com>